Pit Map, markdown for help dialog, & other qol
Previously, in [domain.tld]/api-data, loading data from Nexus loads everything, but does not render the pit map, as shown with "  Visual map rendering will be available in a future update." Now, that map is renderable.

There was also another issue where even if there is a saved event key, the textbox does not autofill. This is now fixed, but I notice that if theres no typing activity to the textbox, then the data pages will not take effect.

Then, I noticed that the help dialog pages (for ex [domain.tld]/qr-data-transfer) would render as: **Consolidated QR Transfer Features:**
 - **Scouting Data Transfer** - Share match scouting data between devices . Clearly, some markdown render was occuring, since it properly showed the bullet and showed the image section, but it wasn't rendering the **. That is now fixed.